# chore: Improve character delimited code performance somewhat