--- a/Cargo.toml
+++ b/Cargo.toml
@@ -253,6 +253,7 @@ logfmt = { version = "0.0.2", default-features = false, optional = true }
 lru = { version = "0.7.1", default-features = false, optional = true }
 maxminddb = { version = "0.21.0", default-features = false, optional = true }
 md-5 = { version = "0.10", optional = true }
+memchr = { version = "2.4", default-features = false, optional = true }
 # make sure to update the external docs when the Lua version changes
 mlua = { version = "0.7.0", default-features = false, features = ["lua54", "send", "vendored"], optional = true }
 mongodb = { version = "2.1.0", default-features = false, features = ["tokio-runtime"], optional = true }
@@ -427,7 +428,7 @@ enrichment-tables = ["enrichment-tables-file"]
 enrichment-tables-file = [ "csv", "seahash", "hash_hasher" ]
 
 # Codecs
-codecs = ["smallvec"]
+codecs = ["smallvec", "memchr"]
 
 # Sources
 sources = ["sources-logs", "sources-metrics"]