--- a/src/codecs/format/bytes.rs
+++ b/src/codecs/format/bytes.rs
@@ -22,7 +22,7 @@ impl BytesDeserializerConfig {
 #[typetag::serde(name = "bytes")]
 impl DeserializerConfig for BytesDeserializerConfig {
     fn build(&self) -> crate::Result<BoxedDeserializer> {
-        Ok(Box::new(BytesDeserializer))
+        Ok(Box::new(BytesDeserializer::new()))
     }
 }
 
@@ -31,19 +31,23 @@ impl DeserializerConfig for BytesDeserializerConfig {
 /// This deserializer can be considered as the no-op action for input where no
 /// further decoding has been specified.
 #[derive(Debug, Clone)]
-pub struct BytesDeserializer;
+pub struct BytesDeserializer {
+    log_schema_message_key: &'static str,
+}
 
 impl BytesDeserializer {
     /// Creates a new `BytesDeserializer`.
-    pub const fn new() -> Self {
-        Self
+    pub fn new() -> Self {
+        Self {
+            log_schema_message_key: log_schema().message_key(),
+        }
     }
 }
 
 impl Deserializer for BytesDeserializer {
     fn parse(&self, bytes: Bytes) -> crate::Result<SmallVec<[Event; 1]>> {
         let mut log = LogEvent::default();
-        log.insert(log_schema().message_key(), bytes);
+        log.insert_flat(self.log_schema_message_key, bytes);
         Ok(smallvec![log.into()])
     }
 }