This commit improves the performance of the character delimited, byte
deserializing codec somewhat. The major changes to observe here are the use of
memchr for searching for our needle in the haystack instead of a custom loop and
using this to discard over-long sub-buffers, tightening up the loop in that
struct. Included is the program codec_perf that replicates the benchmark loop,
allowing for perf collection. As of this commit the major time sinks are:

* `<vector::codecs::format::bytes::BytesDeserializer as Deserializer>::parse`
* `vector::codecs::framing::character_delimited::CharacterDelimitedDecoder::decode`
* `BytesMut::split_to`
* `alloc::collections::btree::map::entry::VacantEntry<K,V>::insert`

This `decode` is the caller of `split_to`, `parse` calls `LogEvent::insert_flat`
which is, itself, a btree insertion. Timings captured on my system:

* `character_delimited_bytes/no_max`: 114MiB/s
* `character_delimited_bytes/no_max`: 209MiB/s
* `newline_bytes/no_max`: 369MiB/s
* `newline_bytes/no_max`: 1.5GiB/s

REF #10543

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>

<!--
**Your PR title must conform to the conventional commit spec!**

  <type>(<scope>)!: <description>

  * `type` = chore, enhancement, feat, fix, docs
  * `!` = OPTIONAL: signals a breaking change
  * `scope` = Optional when `type` is "chore" or "docs", available scopes https://github.com/timberio/vector/blob/master/.github/semantic.yml#L20
  * `description` = short description of the change

Examples:

  * enhancement(file source): Add `sort` option to sort discovered files
  * feat(new source): Initial `statsd` source
  * fix(file source): Fix a bug discovering new files
  * chore(external docs): Clarify `batch_size` option
-->